Make sure dependencies are installed
Found no better way to make Visual Studio install Clide when installing this extension.
Adding Clide to Dependencies doesn't work. VisualStudio fails to resolve it during installation.
We can include 'Clide.Core.Windows.dll' as a Mef component in this extension, but that causes
issues if Xamarin is installed too (two 'Clide.Core.Windows.dll' so components are exported twice).
As such, this extension sadly requires Xamarin to be installed as well for now. It's a quite
bulky dependency, so we should continue looking for possible alternatives to this.