Fix launch issue when GodotProjectDir not same as SolutionDir
If the Solution for a Godot project is not the same as the project directory the extension will
fail to launch the debug target with a MessageBox showing 'No Godot editor instance connected'.

The reason is that the extension takes SolutionDir as the Godot project dir. While this hold true for
simple default Godot projects this must not be the case. In complex projects developers may
restructure their projects into a alternative structure, where the solution may be at another location.

The *.csproj file has the MSBuild property 'GodotProjectDir'. Now the value of this property is fetched
and will be used as the project dir.

Note: Godot itself has some issues with projects varying to the default project. To fix this godot issues,
modifications in the GodotSDK MSBuild props file have to be done.